move safetype
